+++ MK4duo/Configuration_Delta.h
@@ -595,8 +595,13 @@
  *****************************************************************************************/
 // delta homing speeds must be the same on xyz. Homing speeds (mm/m)
 #define HOMING_FEEDRATE_XYZ (100*60)
+
+// Slow Homing feature reduce Acceleration and Jerk only for homing
+//#define SLOW_HOMING
+
 // homing hits the endstop, then retracts by this distance, before it tries to slowly bump again:
 #define XYZ_HOME_BUMP_MM 5
+
 // Re-Bump Speed Divisor (Divides the Homing Feedrate)
 #define XYZ_BUMP_DIVISOR 10
 /*****************************************************************************************/