+++ MK4duo/Configuration_Motor_Driver.h
@@ -72,21 +72,33 @@
 // Select this if use software SPI. Choose pins in Configuration_pins.h
 //#define TMC_USE_SW_SPI
 
-// Use StallGuard2 to sense an obstacle and trigger an endstop.
+// Use StallGuard2 to home / probe X, Y, Z.
+//
+// TMC2130, TMC2160, TMC2660, TMC5130, and TMC5160 only
 // Connect the stepper driver's DIAG1 pin to the X/Y/Z endstop pin.
 // X, Y, and Z homing will always be done in spreadCycle mode.
 //
 // X/Y/Z_STALL_SENSITIVITY is used for tuning the trigger sensitivity.
-// Higher values make the system LESS sensitive.
-// Lower value make the system MORE sensitive.
-// Too low values can lead to false positives, while too high values will collide the axis without triggering.
-// It is advised to set X/Y/Z_HOME_BUMP_MM to 0.
-// M914 X/Y/Z to live tune the setting
-// TMC2130 or TMC5130 only
+// Use M914 X Y Z to live-adjust the sensitivity.
+//  Higher: LESS sensitive. (Too high => failure to trigger)
+//   Lower: MORE sensitive. (Too low  => false positives)
+//
+// It is recommended to set [XYZ]_HOME_BUMP_MM to 0.
+//
+// SPI_ENDSTOPS  *** Beta feature! *** TMC2130 Only ***
+// Poll the driver through SPI to determine load when homing.
+// Removes the need for a wire from DIAG1 to an endstop pin.
+//
+// IMPROVE_HOMING_RELIABILITY tunes acceleration and jerk when
+// homing and adds a guard period for endstop triggering.
+//
 //#define SENSORLESS_HOMING
 #define X_STALL_SENSITIVITY 8
 #define Y_STALL_SENSITIVITY 8
 #define Z_STALL_SENSITIVITY 8
+
+//#define SPI_ENDSTOPS
+//#define IMPROVE_HOMING_RELIABILITY
 
 // Create a 50/50 square wave step pulse optimal for stepper drivers.
 //#define SQUARE_WAVE_STEPPING