+++ MK4duo/Configuration_Version.h
@@ -26,7 +26,7 @@
 #define FIRMWARE_NAME             "MK4duo"
 #define SHORT_BUILD_VERSION       "4.3.9"
 #define BUILD_VERSION             FIRMWARE_NAME "_" SHORT_BUILD_VERSION
-#define STRING_DISTRIBUTION_DATE  "08-08-2019"
+#define STRING_DISTRIBUTION_DATE  "10-08-2019"
 #define FIRMWARE_URL              "marlinkimbra.it"
 
 #endif /* _CONFIGURATION_VERSION_H_ */