+++ MK4duo/src/core/mechanics/cartesian_mechanics.cpp
@@ -198,6 +198,28 @@ void Cartesian_Mechanics::home(const bool homeX/*=false*/, const bool homeY/*=fa
     workspace_plane = PLANE_XY;
   #endif
 
+  // Reduce Acceleration and Jerk for Homing
+  #if ENABLED(SLOW_HOMING) || ENABLED(IMPROVE_HOMING_RELIABILITY)
+    struct slow_homing_t {
+      struct { uint32_t x, y; } acceleration;
+      #if HAS_CLASSIC_JERK
+        struct { float x, y; } jerk;
+      #endif
+    };
+    slow_homing_t slow_homing { 0 };
+    slow_homing.acceleration.x = data.max_acceleration_mm_per_s2[X_AXIS];
+    slow_homing.acceleration.y = data.max_acceleration_mm_per_s2[Y_AXIS];
+    data.max_acceleration_mm_per_s2[X_AXIS] = 100;
+    data.max_acceleration_mm_per_s2[Y_AXIS] = 100;
+    #if HAS_CLASSIC_JERK
+      slow_homing.jerk.x = data.max_jerk[X_AXIS];
+      slow_homing.jerk.y = data.max_jerk[Y_AXIS];
+      data.max_jerk[X_AXIS] = 0;
+      data.max_jerk[Y_AXIS] = 0;
+    #endif
+    planner.reset_acceleration_rates();
+  #endif
+
   // Always home with tool 0 active
   #if HOTENDS > 1
     const uint8_t old_tool_index = tools.extruder.active;
@@ -349,6 +371,16 @@ void Cartesian_Mechanics::home(const bool homeX/*=false*/, const bool homeY/*=fa
   // Restore the active tool after homing
   #if HOTENDS > 1
     tools.change(old_tool_index, true);
+  #endif
+
+  #if ENABLED(SLOW_HOMING) || ENABLED(IMPROVE_HOMING_RELIABILITY)
+    data.max_acceleration_mm_per_s2[X_AXIS] = slow_homing.acceleration.x;
+    data.max_acceleration_mm_per_s2[Y_AXIS] = slow_homing.acceleration.y;
+    #if HAS_CLASSIC_JERK
+      data.max_jerk[X_AXIS] = slow_homing.jerk.x;
+      data.max_jerk[Y_AXIS] = slow_homing.jerk.y;
+    #endif
+    planner.reset_acceleration_rates();
   #endif
 
   lcdui.refresh();