+++ MK4duo/src/core/printer/printer.cpp
@@ -557,6 +557,17 @@ void Printer::stop() {
  */
 void Printer::idle(const bool ignore_stepper_queue/*=false*/) {
 
+  #if ENABLED(SPI_ENDSTOPS)
+    if (endstops.tmc_spi_homing.any
+      #if ENABLED(IMPROVE_HOMING_RELIABILITY)
+        && ELAPSED(millis(), tmc.sg_guard_period)
+      #endif
+    ) {
+      for (uint8_t i = 4; i--;) // Read SGT 4 times per idle loop
+        if (endstops.tmc_spi_homing_check()) break;
+    }
+  #endif
+
   lcdui.update();
 
   #if ENABLED(HOST_KEEPALIVE_FEATURE)