+++ MK4duo/src/lib/macros.h
@@ -24,12 +24,20 @@
 // NoPin
 #define NoPin     -1
 
+// Use NUM_ARGS(__VA_ARGS__) to get the number of variadic arguments
+#define _NUM_ARGS(_0,_24_,_23,_22,_21,_20,_19,_18,_17,_16,_15,_14,_13,_12,_11,_10,_9,_8,_7,_6,_5,_4,_3,_2,_1,N,...) N
+#define NUM_ARGS(V...) _NUM_ARGS(0,V,24,23,22,21,20,19,18,17,16,15,14,13,12,11,10,9,8,7,6,5,4,3,2,1,0)
+
 // Macros for CAT args
-#define _CAT(a,V...) a##V
+#define _CAT(a,V...)  a##V
 
 // Macros to support option testing
 #define ENABLED   defined
 #define DISABLED !defined
+
+// Macros to support pins/buttons exist testing
+#define PIN_EXISTS(PN)    (defined(PN##_PIN) && PN##_PIN > NoPin)
+#define BUTTON_EXISTS(BN) (defined(BTN_##BN) && BTN_##BN > NoPin)
 
 // The axis order in all axis related arrays is X, Y, Z, E
 #define NUM_AXIS  4
@@ -205,10 +213,6 @@
 // Limit an index to an array size
 #define ALIM(I,ARR)               MIN(I, COUNT(ARR) - 1)
 
-// Pin Exist
-#define PIN_EXISTS(PN)            (defined(PN##_PIN) && PN##_PIN > NoPin)
-#define BUTTON_EXISTS(BN)         (defined(BTN_## BN) && BTN_## BN >= 0)
-
 #define NOOP                      (void(0))
 
 #define CEILING(x,y)              (((x) + (y) - 1) / (y))
@@ -318,10 +322,6 @@
   }
 
 #else
-
-  // Pass NUM_ARGS(__VA_ARGS__) to use the number of arguments
-  #define _NUM_ARGS(_0,_24_,_23,_22,_21,_20,_19,_18,_17,_16,_15,_14,_13,_12,_11,_10,_9,_8,_7,_6,_5,_4,_3,_2,_1,N,...) N
-  #define NUM_ARGS(...) _NUM_ARGS(0, __VA_ARGS__ ,24,23,22,21,20,19,18,17,16,15,14,13,12,11,10,9,8,7,6,5,4,3,2,1,0)
 
   #define MIN_2(a,b)      ((a)<(b)?(a):(b))
   #define MIN_3(a,...)    MIN_2(a,MIN_2(__VA_ARGS__))